src/ml-service/packages/ml-api/requirements.txt
# api
flask==1.1.1

#production requirements
numpy==1.17.4
scikit-learn==0.21.3
pandas==0.25.3
statsmodels==0.10.1
matplotlib==3.1.1
psutil==5.6.5

# packaging
setuptools==41.6.0
wheel==0.32.3

#schema validation
marshmallow== 3.2.2
geopy==1.20.0

# testing requirements
pytest>=4.6.6,<5.0.0
#environment setup
gunicorn==20.0.4

#google cloud services
google-cloud-bigquery==1.22.0
google-python-cloud-debugger